Refactor, move Appbar in separate widget
