+++ lib/src/ui/posts/posts_screen.dart
@@ -1,4 +1,4 @@
-import 'package:daily_mobile_app/src/ui/common/appbar_logo.dart';
+import 'package:daily_mobile_app/src/ui/common/posts_appbar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
@@ -11,11 +11,7 @@ class _PostsPageState extends State<PostsPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        automaticallyImplyLeading: false,
-        centerTitle: true,
-        title: AppBarLogo(),
-      ),
+      appBar: PostsAppBar(),
       body: Container(),
     );
     ;
